Add support for scraping player stats per-map
The script now pulls the statistics for each player on every map for
each match, i.e. a BO3 with 3 maps will pull 30 rows of data, 3 for
each player. The error trap for HTTP errors is now located in the html
method and returns None for any errors, which must be handled when
calling getHTML().